+++ packages/platform-shared/src/node/index.ts
@@ -2,4 +2,3 @@ export * from './platform/runtimeFiles';
 export { sharedPlugin } from './platform/plugin';
 export * from '../runtime/runtimeConfig';
 export * from './route';
-export * from './route-layout';
