--- packages/platform-shared/src/node/route-layout/index.ts
@@ -1,5 +1,0 @@
-export {
-  getPageAndLayoutRoutes,
-  getApiRoutes,
-  getMiddlewareRoutes
-} from './route';
