+++ packages/platform-shared/src/node/route/__tests__/parseDynamicPath.test.ts
@@ -1,4 +1,4 @@
-import parseDynamicPath from '../parseDynamicPath';
+import { parseDynamicPath } from '../helpers';
 
 describe('parse dynamic path', () => {
   test('should work', async () => {
