+++ packages/platform-shared/src/node/route/route.ts
@@ -1,189 +1,323 @@
-import { join, relative } from 'path';
-import { IUserRouteConfig } from '@shuvi/service';
-import { IRouteRecord } from '@shuvi/router';
+import { basename, extname, join, dirname } from 'path';
+import {
+  getAllowFilesAndDirs,
+  hasAllowFiles,
+  normalize,
+  SupportFileType
+} from './helpers';
 
-import parseDynamicPath from './parseDynamicPath';
+import type { IRouteRecord } from '@shuvi/router';
+import { isDirectory } from '@shuvi/utils/lib/file';
 
-export type SubscribeFn = (v: IRouteRecord[]) => void;
-
-const jsExtensions = ['js', 'jsx', 'ts', 'tsx'];
-
-const RouteFileRegExp = new RegExp(`\\.(?:${jsExtensions.join('|')})$`);
-
-const MIDDLEWAREJS = '_middleware.js';
-const MIDDLEWARETS = '_middleware.ts';
-
-function getMiddlewareFile(fileToTransform: IFilesObject) {
-  if (fileToTransform[MIDDLEWAREJS]) {
-    return MIDDLEWAREJS;
-  }
-  if (fileToTransform[MIDDLEWARETS]) {
-    return MIDDLEWARETS;
-  }
-  return null;
+interface RawFileRoute {
+  type: SupportFileType;
+  segment: string;
+  name: string;
+  filepath: string;
 }
 
-function isLayout(filepath: string) {
-  return filepath.endsWith('/_layout');
+interface RawDirRoute {
+  type: 'dir';
+  filepath: string;
+  segment: string;
+  parentSegment: string;
+  children: (RawFileRoute | RawDirRoute)[];
 }
 
-interface IFilesObject {
-  [file: string]: IFilesObject;
-}
+type RawRoute = RawFileRoute | RawDirRoute;
 
-/**
- * This method transform files into a nested hash object of each file level.`
- * Example ['/a/1', '/a/2', '/b/1'] will become
- * { a: { 1: {}, 2: {} }, b: { 1: {} } }
- * */
-const transformFilesObject = (
-  files: string[],
-  ignoreLayout: boolean
-): IFilesObject => {
-  return files.reduce((acc, file) => {
-    file
-      .split('/')
-      .filter(Boolean)
-      .forEach((route, index, arr) => {
-        let objToTraverse = acc;
-
-        if (
-          index === 0 &&
-          !ignoreLayout &&
-          normalizeFilePath(route) === '/_layout'
-        ) {
-          console.warn(
-            'Top level _layout is not supported and will be ignored.'
-          );
-        } else {
-          for (let i = 0; i < index; i++) {
-            objToTraverse = objToTraverse[arr[i]];
-          }
-
-          if (objToTraverse && objToTraverse[route]) {
-            objToTraverse[route] = {
-              ...objToTraverse[route]
-            };
-          } else {
-            objToTraverse[route] = {};
-          }
-        }
-      });
-    return acc;
-  }, {} as IFilesObject);
+export type RouteException = {
+  type: SupportFileType | 'dir';
+  msg: string;
 };
 
-export function normalizeFilePath(filepath: string) {
-  const res = filepath
-    // Remove the file extension from the end
-    .replace(/\.\w+$/, '')
-    // Convert to unix path
-    .replace(/\\/g, '/');
+export const getRawRoutesFromDir = async (dirname: string) => {
+  const warnings: RouteException[] = [];
+  const errors: RouteException[] = [];
+  const _getRawRoutesFromDir = async (
+    dirname: string,
+    rawRoutes: RawRoute[],
+    segment: string
+  ) => {
+    const files = await getAllowFilesAndDirs(dirname);
+    const onlyHasDir = !hasAllowFiles(files);
 
-  return res.charAt(0) !== '/' ? '/' + res : res;
-}
+    if (!files.length) {
+      warnings.push({
+        type: 'dir',
+        msg: `${dirname} is empty dir!`
+      });
+    }
 
-export function normalizeRoutePath(rawPath: string) {
-  // /xxxx/index -> /xxxx/
-  let routePath = rawPath.replace(/\/index$/, '/');
+    for (const file of files) {
+      const filepath = join(dirname, file);
+      const isDir = await isDirectory(filepath);
 
-  // remove the last slash
-  // e.g. /abc/ -> /abc
-  if (routePath !== '/' && routePath.slice(-1) === '/') {
-    routePath = routePath.slice(0, -1);
-  }
+      if (isDir) {
+        if (onlyHasDir) {
+          // only indent segment,routes was in same level.
+          await _getRawRoutesFromDir(filepath, rawRoutes, `${segment}/${file}`);
+          continue;
+        }
 
-  routePath = parseDynamicPath(routePath);
+        const rawRoute: RawDirRoute = {
+          type: 'dir',
+          filepath,
+          segment: normalize(file).replace(/^\//, ''),
+          parentSegment: normalize(segment).replace(/^\//, ''),
+          children: []
+        };
 
-  return routePath;
-}
+        await _getRawRoutesFromDir(filepath, rawRoute.children, file);
+        rawRoutes.push(rawRoute);
 
-function filterRouteFile(name: string) {
-  if (name.charAt(0) === '.') return false;
+        continue;
+      }
+      const ext = extname(file);
+      const type = basename(file, ext) as SupportFileType;
+      rawRoutes.push({
+        name: file,
+        segment: normalize(segment),
+        filepath,
+        type
+      });
+    }
+    return rawRoutes;
+  };
+  const rawRoutes = await _getRawRoutesFromDir(dirname, [], '');
+  return {
+    rawRoutes,
+    warnings,
+    errors
+  };
+};
 
-  return RouteFileRegExp.test(name);
-}
+export const getApiRoutes = async (dir: string) => {
+  const getConflictWaring = (
+    rawRoute: RawRoute,
+    conflictRawRoute: RawRoute
+  ) => {
+    return `Find both ${basename(conflictRawRoute.filepath)} and ${basename(
+      rawRoute.filepath
+    )} in "${dirname(rawRoute.filepath)}"!, only "${basename(
+      conflictRawRoute.filepath
+    )}" is used.`;
+  };
 
-export const generateRoute = (
-  fileToTransform: IFilesObject,
-  pageDirectory: string,
-  ignoreLayout: boolean,
-  middlewareArr: string[]
-) => {
-  const routes: IRouteRecord[] = [];
-  const middlewareFile = getMiddlewareFile(fileToTransform);
-  if (middlewareFile) {
-    delete fileToTransform[middlewareFile];
-    const middlewarePath = join(pageDirectory, middlewareFile);
-    middlewareArr.push(middlewarePath);
-  }
-  Object.entries(fileToTransform).forEach(([fileName, nestedRoute], _, arr) => {
-    let route: IRouteRecord;
-    let routePath = normalizeRoutePath(normalizeFilePath(fileName));
-    const isDirectory = Object.values(nestedRoute).length > 0;
-    route = {
-      path: routePath,
-      middlewares: middlewareArr
-    } as IRouteRecord;
+  const { rawRoutes, warnings, errors } = await getRawRoutesFromDir(dir);
+  const allowTypes = ['api', 'dir'];
 
-    // if a directory have _layout, treat it as its own source
-    if (isDirectory) {
-      if (!ignoreLayout) {
-        const layoutFile = Object.keys(nestedRoute).find(route =>
-          isLayout(normalizeRoutePath(normalizeFilePath(route)))
+  const _getApiRoutes = (
+    rawRoutes: RawRoute[],
+    routes: IRouteRecord[],
+    prefix: string = ''
+  ) => {
+    const page = rawRoutes.find(route => route.type === 'page');
+    const layout = rawRoutes.find(route => route.type === 'layout');
+    const allowedRoutes = rawRoutes.filter(route =>
+      allowTypes.includes(route.type)
+    );
+
+    for (let rawRoute of allowedRoutes) {
+      prefix = prefix === '/' ? '' : prefix;
+
+      if (rawRoute.type === 'dir') {
+        _getApiRoutes(
+          rawRoute.children,
+          routes,
+          prefix + '/' + rawRoute.parentSegment
         );
-        if (layoutFile) {
-          route.filepath = join(pageDirectory, fileName, layoutFile);
-          // delete _layout
-          delete nestedRoute[layoutFile];
+        continue;
+      }
+      if (rawRoute.type === 'api') {
+        if (layout) {
+          warnings.push({
+            type: 'api',
+            msg: getConflictWaring(rawRoute, layout)
+          });
+          continue;
+        }
+
+        if (page) {
+          warnings.push({
+            type: 'api',
+            msg: getConflictWaring(rawRoute, page)
+          });
+          continue;
+        }
+
+        let path = prefix + '/' + rawRoute.segment;
+        if (path === '//') {
+          path = '/';
+        }
+
+        routes.push({
+          path,
+          filepath: rawRoute.filepath
+        });
+      }
+    }
+    return routes;
+  };
+
+  const routes = _getApiRoutes(rawRoutes, [], '');
+  const filterException = (e: RouteException) => e.type === 'api';
+
+  return {
+    routes,
+    warnings: warnings.filter(filterException),
+    errors: errors.filter(filterException)
+  };
+};
+
+export const getPageAndLayoutRoutes = async (dirname: string) => {
+  const { rawRoutes, warnings, errors } = await getRawRoutesFromDir(dirname);
+  const allowTypes = ['dir', 'page', 'layout'];
+
+  const _getPageAndLayoutRoutes = async (
+    rawRoutes: RawRoute[],
+    routes: IRouteRecord[],
+    segment = ''
+  ) => {
+    const allowedRawRoutes = rawRoutes.filter(route =>
+      allowTypes.includes(route.type)
+    );
+    const hasLayout = allowedRawRoutes.some(route => route.type === 'layout');
+
+    const route = {} as IRouteRecord;
+
+    if (hasLayout) {
+      route.children = [];
+    }
+
+    for (const rawRoute of allowedRawRoutes) {
+      if (rawRoute.type === 'page' && hasLayout) {
+        route.children!.push({
+          path: '',
+          component: rawRoute.filepath
+        });
+        continue;
+      }
+
+      if (rawRoute.type === 'layout' || rawRoute.type === 'page') {
+        let prefix = segment;
+        let suffix = rawRoute.segment;
+
+        route.component = rawRoute.filepath;
+        route.path = `${prefix}${suffix}`;
+
+        routes.push({
+          ...route
+        });
+        continue;
+      }
+
+      if (rawRoute.type === 'dir') {
+        const workRoutes = hasLayout ? route.children! : routes;
+        const preSegment = hasLayout
+          ? ''
+          : `${segment}${rawRoute.parentSegment}/`;
+        await _getPageAndLayoutRoutes(
+          rawRoute.children,
+          workRoutes,
+          preSegment
+        );
+      }
+    }
+
+    return routes;
+  };
+
+  const routes = await _getPageAndLayoutRoutes(rawRoutes, []);
+
+  routes.forEach(route => {
+    if (!route.path.startsWith('/')) {
+      route.path = `/${route.path}`;
+    }
+  });
+
+  return {
+    routes,
+    warnings: warnings.filter(warning => allowTypes.includes(warning.type)),
+    errors: errors.filter(error => allowTypes.includes(error.type))
+  };
+};
+
+export type MiddlewareRecord = {
+  middlewares: string[];
+  path: string;
+  children?: MiddlewareRecord[];
+};
+
+export const getMiddlewareRoutes = async (dirname: string) => {
+  const { rawRoutes, warnings, errors } = await getRawRoutesFromDir(dirname);
+
+  const _getMiddlewareRoutes = async (
+    rawRoutes: RawRoute[],
+    routes: MiddlewareRecord[],
+    segment: string,
+    parentMiddlewares: string[]
+  ) => {
+    segment = segment === '/' ? '' : segment;
+
+    const currentLevelMiddleware = rawRoutes.find(
+      route => route.type === 'middleware'
+    );
+
+    const middlewares = currentLevelMiddleware
+      ? [...parentMiddlewares, currentLevelMiddleware.filepath]
+      : parentMiddlewares;
+
+    const hasLayout = rawRoutes.some(route => route.type === 'layout');
+    const hasPage = rawRoutes.some(route => route.type === 'page');
+
+    for (const rawRoute of rawRoutes) {
+      if (rawRoute.type === 'dir') {
+        await _getMiddlewareRoutes(
+          rawRoute.children,
+          routes,
+          //segment + '/' + (rawRoute.parentSegment || rawRoute.segment),
+          segment + '/' + rawRoute.parentSegment,
+          middlewares
+        );
+        continue;
+      }
+
+      const isPage = rawRoute.type === 'page';
+      const isLayout = rawRoute.type === 'layout';
+      const isMiddleware = rawRoute.type === 'middleware';
+      const onlyHasPage = isPage && !hasLayout;
+      const onlyHasMiddleware = isMiddleware && !hasLayout && !hasPage;
+
+      if (isLayout || onlyHasPage || onlyHasMiddleware) {
+        if (middlewares.length) {
+          let path = segment + '/' + rawRoute.segment;
+          if (path === '//') {
+            path = '/';
+          }
+          routes.push({
+            path,
+            middlewares
+          });
         }
       }
-      route.children = generateRoute(
-        nestedRoute,
-        join(pageDirectory, fileName),
-        ignoreLayout,
-        middlewareArr.slice()
-      ); // inner directory
-    } else {
-      route.filepath = join(pageDirectory, fileName);
     }
-    routes.push(route);
+
+    return routes;
+  };
+
+  const routes = await _getMiddlewareRoutes(rawRoutes, [], '', []);
+  routes.forEach(route => {
+    if (!route.path.startsWith('/')) {
+      route.path = `/${route.path}`;
+    }
   });
-  return routes;
+
+  const exceptionFilter = (e: RouteException) => e.type === 'middleware';
+
+  return {
+    routes,
+    warnings: warnings.filter(exceptionFilter),
+    errors: errors.filter(exceptionFilter)
+  };
 };
-
-export const getRoutesFromFiles = (
-  files: string[],
-  filesDir: string,
-  ignoreLayout: boolean = false
-): IRouteRecord[] => {
-  const filteredFiles = files
-    .map(file => relative(filesDir, file))
-    .filter(filterRouteFile);
-  const transformedFiles = transformFilesObject(filteredFiles, ignoreLayout);
-  return renameFilepathToComponent(
-    generateRoute(transformedFiles, filesDir, ignoreLayout, [])
-  );
-};
-
-export function renameFilepathToComponent(
-  routes: IRouteRecord[]
-): IUserRouteConfig[] {
-  const res: IUserRouteConfig[] = [];
-  for (let index = 0; index < routes.length; index++) {
-    const { path, filepath, children } = routes[index];
-    const route = {
-      path
-    } as IUserRouteConfig;
-
-    if (filepath) {
-      route.component = filepath;
-    }
-
-    if (children && children.length > 0) {
-      route.children = renameFilepathToComponent(children);
-    }
-    res.push(route);
-  }
-  return res;
-}
