+++ packages/platform-web/src/lib/features/html-render/index.ts
@@ -18,7 +18,7 @@ import {
 import server from './server-plugin-custom-server';
 import { FileOptions } from '@shuvi/service/lib/project';
 import { isDirectory } from '@shuvi/utils/lib/file';
-import { getPageAndLayoutRoutes } from '@shuvi/platform-shared/lib/node/route-layout/route';
+import { getPageAndLayoutRoutes } from '@shuvi/platform-shared/lib/node';
 export { IRenderToHTML } from './hooks';
 export {
   getSSRMiddleware,
